Fix the menu positioning
